feat(prelaunch): close roadmap §2.1 + 2.2 + 2.3 + 2.4 + 2.5 + 3.6 + pool-exhausted bug
Five critical pre-launch gaps closed in one bundle. The proto + Flyway
migrations interlock across services so they must land together.

§2.1 Auth — real signup/login/JWT verification (api-gateway):
  * Argon2id password hash (Spring Security crypto + BouncyCastle)
  * Endpoints: POST /api/v1/auth/{signup,login,refresh,logout}
  * Refresh-token rotation; only SHA-256(token) stored server-side
  * JWT now carries kid header for key rotation; typ=access vs refresh
  * Dev mint POST /api/v1/auth/token DELETED
  * SecurityConfig: everything except /actuator/health + /auth/{signup,login,refresh}
    requires a valid access token
  * V5__auth.sql adds password_hash, refresh_tokens, auth_events
  * AuthServiceTest with 7 happy/edge tests

§2.2 Schema discipline — one DB, one migration tool:
  * Every service connects to `onlinejudge` (api-gateway, execution-worker,
    problem-service, contest-service-dev-yml)
  * V3__unify_schema.sql idempotently adds problems / test_cases / users
    to onlinejudge so Flyway can land them once api-gateway flips
  * problem-service entity widened int→long (CRDB INT is BIGINT)
  * SPRING_JPA_HIBERNATE_DDL_AUTO=none override DROPPED — `validate` is
    the right end state now that entity types match
  * init.sql replaced with OBSOLETE banner; mount removed from compose;
    terraform stops base64-injecting it
  * infra/scripts/db-migration-onlinejudge.sql — operator runbook that
    DROPs orphan onlinejudge.{submissions,outbox_events,idempotency_keys}
    (wrong shape from old init.sql path) then INSERT…SELECT-copies the
    live defaultdb rows over after Flyway lands

§2.3 Per-problem time/memory limits propagated to the agent:
  * problem-service /test-cases response now wraps the array as
    TestCaseBundle { time_limit_ms, memory_limit_mb, test_cases[] }
  * ProblemServiceClient + TestCaseFetcher.ProblemSpec carry limits
  * FirecrackerExecutionService passes time → AgentClient (per-test
    wall clock) and memory → SandboxManagerClient (lease) — both
    already honour them in the SM cgroup applier + Go agent runTimeout
  * SubmissionConsumer.determineVerdict now reads per-problem limit
    instead of literal 2000 ms

§2.4 VerdictEvent per-ordinal breakdown:
  * Added repeated PerTestVerdict per_test = 13 to VerdictEvent
    (proto3-safe; downstream consumers that don't regenerate just ignore)
  * SubmissionConsumer copies AgentClient.PerTestResult → PerTestVerdict
    (ordinal, verdict, time_ms, memory_kb, stdout_hash — NO stdout bytes,
    preserves test-case secrecy)

§2.5 Idempotency reclaim cap + DLQ:
  * idempotency_keys.attempts INT NOT NULL DEFAULT 0 (V4 Flyway)
  * INSERT path sets attempts = 1; reclaim path SETs attempts+1 with
    `attempts < :maxAttempts` guard
  * Above the cap, claim returns POISON; SubmissionConsumer publishes a
    JSON DLQ envelope to ${app.kafka.topic.dlq:submissions.dlq} and acks
  * markPoison() + releaseClaim() (for transient backpressure cleanup)
  * Claim moved AFTER successful sandbox lease so a 503 pool_exhausted
    doesn't burn an attempt

§3.6 Source-code size cap:
  * SubmissionRequest.code @Size(max=65536); Tomcat-level body caps to
    match; pre-Jackson Content-Length check returns 413 cleanly

Pool-exhausted bug surfaced in last smoke:
  * FirecrackerExecutionService now distinguishes a 503 pool_exhausted
    from a real runtime error — throws PoolExhaustedException carrying
    retry_after_ms
  * SubmissionConsumer.processSubmission catches it separately,
    releaseClaim()s the idempotency row, and ack.nack(retry_after) so
    the message redelivers without counting against the attempts cap

DB migration order (operator runbook lives at the top of
infra/scripts/db-migration-onlinejudge.sql):
  1. Stop api-gateway. 2. Drop orphan tables. 3. Push new compose +
     start api-gateway → Flyway V1..V5 run on `onlinejudge`. 4. Copy
     live data from defaultdb → onlinejudge. 5. Bounce compute VM.

Build / test:
  * gradle :common:build (proto regen) + :execution-worker:test +
    :problem-service:test + :api-gateway:test all green.
  * Images that need rebuild + AR push: api-gateway, problem-service,
    execution-worker. Rootfs unchanged (agent main.go already read
    time_limit_ms from the request).

Coordination notes:
  * Agent W's V3 idempotency migration renamed to V4 to avoid version
    collision with Agent S's V3 unify-schema.
  * DockerExecutionService's wrapper ExecutionResult class removed
    (it shadowed ExecutionBackend.ExecutionResult and broke @Override);
    DockerExecutionServiceTest migrated to the interface record.
  * IdempotencyServiceTest's `List.of(new Object[]{...})` varargs
    gotcha (spreads into N elements instead of singleton list) replaced
    with Collections.singletonList — 5 test cases were silently
    constructing wrong-shaped result lists.

+++ problem-service/src/main/java/com/onlinejudge/problem/dto/ProblemDto.java
@@ -17,6 +17,15 @@ public record ProblemDto(
         int points
 ) {
     public static ProblemDto from(Problem p) {
-        return new ProblemDto(p.getId(), p.getTitle(), p.getTimeLimitMs(), p.getMemoryLimitMb(), p.getPoints());
+        // Entity fields are `long` (matching CRDB's BIGINT-flavoured `INT` so
+        // Hibernate's ddl-auto=validate succeeds). Real values are bounded
+        // (time_limit_ms < 60_000, memory_limit_mb < 16_384, points < 1_000_000)
+        // so a narrowing cast on the way out is safe.
+        return new ProblemDto(
+                p.getId(),
+                p.getTitle(),
+                Math.toIntExact(p.getTimeLimitMs()),
+                Math.toIntExact(p.getMemoryLimitMb()),
+                Math.toIntExact(p.getPoints()));
     }
 }
